requirements.txt
markdown>=3.5
Jinja2>=3.1
PyYAML>=6.0
Pygments>=2.17
